fix: add same sorting to web and small polishing
