+++ apps/desktop/src/lib/components/shared/command-menu/helpers.ts
@@ -36,19 +36,3 @@ export const getAllItems = async (
 
 	return items;
 };
-
-export const sortFileEntry = (a: FileEntry, b: FileEntry): number => {
-	const isDirectory = (file: FileEntry) => file.children != null;
-
-	if (isDirectory(a) && isDirectory(b)) {
-		return naturalSort(a.name!, b.name!);
-	}
-	if (isDirectory(a)) return -1;
-	if (isDirectory(b)) return 1;
-
-	return naturalSort(a.name!, b.name!);
-};
-
-const naturalSort = (a: string, b: string): number => {
-	return a.localeCompare(b, undefined, { numeric: true, sensitivity: 'base' });
-};
